ai_server/requirements.txt
flask
transformers
sentence-transformers
torch
openai-whisper
ffmpeg-python
requests
